Adds debug manager class with OTA updates
Implements a debug mode that can be activated via a pin.

This mode allows for over-the-air (OTA) updates.  A new
`DebugManager` class handles the debug mode logic, including
checking the debug pin, connecting to WiFi, starting OTA, and
handling OTA updates. It also provides a method to enter deep
sleep if not in debug mode.

Includes ArduinoJson for serializing sensor data.
Adds forward declarations for `PubSubClient` to resolve
dependency issues.